Add base config for rollup
